bootstraper/requirements.txt
# --- Critical Version Alignment ---
# Pinning these three is the ONLY way to satisfy the s3fs <-> pyiceberg conflict
fsspec==2023.10.0
s3fs==2023.10.0
pyarrow>=14.0.0

# --- Core Logic ---
# Note: s3fs is already included in the [s3fs] extra
pyiceberg[sql-postgres,s3fs,pyarrow]==0.6.0
pandas==2.0.3
numpy>=1.24.0,<2.0.0
faiss-cpu>=1.8.0

# --- ML & Utilities ---
sentence-transformers>=2.5.0
psycopg2-binary>=2.9.9
requests>=2.31.0
aiobotocore==2.7.0
botocore==1.31.64
boto3==1.28.64
python-dotenv>=1.0.1
pyyaml>=6.0.1
tqdm>=4.66.0
kaggle>=1.6.0